update imports + small fixes
