+++ src/styles/nprogress-custom.scss
@@ -10,7 +10,7 @@
   left: 0;
 
   width: 100%;
-  height: 3px;
+  height: 8px;
   border-radius: 2px;
 }
