--- src/types/PostCategory.ts
@@ -1,7 +1,0 @@
-export enum PostCategory {
-  All = '',
-  Uncategorized = 'uncategorized',
-  JavaScript = 'javascript',
-  ReactJS = 'reactjs',
-  NextJS = 'nextjs'
-}
